Move back to legacy netchan hook approach
Using ConVars or global vars to check whether to log or not turned out to be an issue on the 'sendto' function as this is multithreaded, and after big chunks of data (e.g. the baseline snapshot) has been transmitted the channel becomes stuck as it runs out of sync, or crashes as it calls 'NET_EncryptData' without valid data to be encrypted. and the compiler 'optimizes' the function in such way that the output doesn't become what i like to see.. using LVMM clang resulted in more or less the same output. The legacy approach hooks the functions while the process is already running, and unhooks them if logging is no longer needed.